+++ src/nbgrader_jupyterquiz/__main__.py
@@ -1,9 +1,0 @@
-#!/usr/bin/env python
-
-"""Package entry point."""
-
-from nbgrader_jupyterquiz.interface.command_line import main
-
-
-if __name__ == "__main__":  # pragma: no cover
-    main()  # pylint: disable=no-value-for-parameter